Extract edge runtime-incompatible instrumentation
