+++ src/instrumentation.node.ts
@@ -1,0 +1,20 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import * as Sentry from "@sentry/nextjs";
+
+export function register() {
+  Sentry.init({
+    environment: process.env.APP_ENV,
+    dsn: process.env.SENTRY_DSN,
+
+    // Adjust this value in production, or use tracesSampler for greater control
+    tracesSampleRate: ["local", "heroku"].includes(process.env.NODE_ENV)
+      ? 1.0
+      : 0.1,
+
+    // Setting this option to true will print useful information to the console while you're setting up Sentry.
+    debug: false,
+  });
+}
